requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
google-auth==2.29.0
google-api-python-client==2.128.0
